requirements.txt
anthropic>=0.34.0
pdf2image>=1.16.3
Pillow>=10.0.0
opencv-python>=4.8.0,<4.9
numpy>=1.24.0,<2.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
flask>=2.3.0
flask-cors>=4.0.0

